added top 6 projects to the Project component
