+++ src/App.js
@@ -1,6 +1,6 @@
 // Import dependencies
 import React from 'react';
-import Project from './components/Project';
+import Project from './components/Portfolio';
 import './assets/css/App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
