--- src/components/Project/projects.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": "1",
-    "title": "Travel Lovers",
-    "description": "Professional portfolio to showcase skills to employers",
-    "image": "../../assets/images/projectImages/travel-lovers.png",
-    "github": "https://github.com/nicnolen/Travel-Lovers",
-    "deployed": "https://travel--lovers.herokuapp.com/",
-    "tools": "Handlebars.js, Sequelize, SASS, bcrypt"
-  },
-  {
-    "id": "2",
-    "title": "Travel Lovers",
-    "description": "Professional portfolio to showcase skills to employers",
-    "image": "../../assets/images/projectImages/travel-lovers.png",
-    "github": "https://github.com/nicnolen/Travel-Lovers",
-    "deployed": "https://travel--lovers.herokuapp.com/",
-    "tools": "Handlebars.js, Sequelize, SASS, bcrypt"
-  }
-]
